rename e2e_new test directory and build tag to e2e
