+++ test/e2e/mcp_handshake_test.go
@@ -1,4 +1,4 @@
-//go:build e2e_new
+//go:build e2e
 
 /*
 Copyright 2026 The Kubernetes Authors
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2e_new
+package e2e
 
 import (
 	"context"
@@ -31,7 +31,7 @@ import (
 	"sigs.k8s.io/e2e-framework/pkg/envconf"
 	"sigs.k8s.io/e2e-framework/pkg/features"
 
-	f "github.com/kubernetes-sigs/mcp-lifecycle-operator/test/e2e_new/framework"
+	f "github.com/kubernetes-sigs/mcp-lifecycle-operator/test/e2e/framework"
 )
 
 func TestMCPHandshake(t *testing.T) {
